refactor(tui): isolate ambient life regressions
Move the existing ambient-life test module byte-for-byte in behavior behind its owned tests.rs seam. Keep the production owner below the source-structure threshold without changing rendering, accounting, animation, or collision behavior.
